✨ Feat: Preemptively trigger login after google token expired (#212)
* ✨ Feat(Backend): `/api/auth/session/gauth/verify` route

Implement a new api route to validate a user's google oauth session

* 🐛 Fix(Web): Handle google oauth session validation

* refactor(backend,web): make uri to /auth/google

also renames controller method from verifyGAuthSession to verifyGToken to more accurately represent what's happening. Compass doesn't maintain a persistent session with Google like it does with it's user's session (via Supertokens). Instead, it just passes the access token in requests

* refactor(web): convert GoogleOAuthSession to a function

making it a class doesn't add much value, as we're not taking advantage of any class benefits (inheritance, class methods, encapsulation)

* chore(web): Rename `valid`, always set value to boolean
- Rename `valid` to `isValid` to make it more consistent with codebase
- Always set value to either `true` or `false` for consistency

* chore(web): Add return type to `getGAuthClientForUser`

* chore(backend): Handle potential google API errors

* chore(web): change `valid` to `isValid`

Followup to 86beb9f8da8d177099f49e7ecc6768d7f67d05f0

* chore: remove unnecessary typing in google.auth.service

This same type is returned as part of type inference. Explicitly setting the function return type is acceptable is it is somehow different than the inferred type. In this case, the extra code doesn't provide any value and just requires more maintenance.

* chore: remove unnecessary google error handling

* chore(web): create a constant for gauth session failure reason

This'll make it easier to test, access across other files, and rename it when things change

* chore(web): extract response into dedicated type for reusability

keeping it inside `gauth.util.ts` until I figure out where to put it

* chore(web): Use dedicated constant instead of magic string

* chore(web): refactor session effect to async await

purpose of using await statements is to avoid creating traps for future devs when working on this part of the code.

* chore(backend): Implement dedicated error object for google auth session handling

* enhancment(web): rewrite google session expired message

Users don't care what a session is or that it's the Google connection that's requiring us to make them reauthenticate

* chore: convert error toast warning

* chore(backend): remove unused accessToken variable

* chore: update result arg in NoGAuthAccessToken error

This makes it more clear what happened as a result of the error

* refactor: use AuthApi (axios) and use response type for google access token endpoint

* feat(web): extract auth checking from ProtectedRoute into separate hook and

* chore(web): simplify error log in UserContext

* refactor(web): remove auth checking from Login

* chore(web): remove search param logic from Login

* fix(web): don't blindly trigger signout after 401 response

401 is now a valid response to the /api/auth/google endpoint, which is used to validate a google token

* fix(web): explicitly check for false in useAuthCheck

this is needed to distinguish between boolean values (when we know something happened) and the default value of null (when nothing happened). Without this, an infinite loop of routing and auth checks resulted

* fix(web): redirect after clicking login button (no auth) if user already authenticated

This is a compromise between UX and functionality that lets us provide smooth routing while also ensuring that the user doesn't re-authenticated unnecessarily. It does so by checking if the user is already authenticated as a side effect during the Login page load. If so, when the user clicks the Login button they will be redirected. If not, the user will go through the regular login flow and be redirected afterwards

* chore(web): remove console.log in auth.api

---------

Co-authored-by: Tyler Dane <tyler@switchback.tech>