✨ Feat: Update event cursor css when dragging (#268)
* bug(web): remove `cursor: grabbing` from event styles

`cursor` style is overridden in on hover pseudo class, making this line redundant.

* feature(web): Update event cursor css when dragging

* chore(web): rename `useGridClick` to `useGridMouseUp`

Better emphasizes its purpose, since the hook only handles mouseup events.

* feature(web): Implement mouse hold delay for drafting an event

This is a small enhancement that mimics how Google Calendar delays drafting an event when we mouse hold over it without moving the mouse

* chore(web): small refactor

* feature(web): mouse hold delay feature, support all day events

* feature(web): mouse hold delay feature, support draft events

* ♻️ refactor(web): Remove constant from layout and inline in hook

layout.constants is meant for detailing the page's layout, which'll come in handy as we make the page more dynamic and responsive. The values are shared across many parts of the app, so have a dedicated file for it is helpful.

This value isn't related to the layout, and is only used in one place, so it can live there instead.

---------

Co-authored-by: Tyler Dane <tyler@switchback.tech>