bug(web): Fix all day events overlapping
Fixes issue where if an all day event's start and end date is the same date, the event would overlap with another created all day event if they share same date range

The issue was, when we create an all day event with same start-end date, `eventDays` is always an array of a single integer.

So when we pop the array, we empty it and it becomes an empty array, which is not the intended behavior.